Sync with colcon_clean (#47)
* Update deb platforms for release
Added:
* Ubuntu Noble (24.04 LTS pre-release)
* Debian Trixie (testing)
Dropped:
* None

Retained:
* Debian Bookworm (stable)
* Ubuntu Jammy (22.04 LTS)
- https://github.com/colcon/colcon-clean/pull/43

* Update publish-python platforms.
- https://github.com/colcon/colcon-clean/pull/43

* Set long_description_content_type in setup.cfg
The default is assumed to be rst, and PyPI is unhappy that it can't
render it as such. Explicitly stating that the content is markdown
resolves the issue.
- https://github.com/colcon/colcon-clean/pull/33

* Update target platforms for debs, python requirement
- https://github.com/colcon/colcon-clean/commit/c7ff24c1869ac13d022ad02f2dfb2f0b2c87e196

* Add colcon-defaults to test extras_require
to set default test-result-base base path

- https://github.com/colcon/colcon-clean/pull/28

* Add markers to linter tests, move linter imports
Test markers can be used to easily (de-)select tests, and colcon exposes
mechanisms to do so. Linters are a category of tests that are commonly
called out.

Additionally, if we move the imports for some of our single-purpose
tests into the test function, we can avoid installing the linter
dependencies entirely. This is a common case in platform packaging, where
linter errors are not actionable and the dependencies are not typically
installed.

- https://github.com/colcon/colcon-clean/commit/1ba89236e86184b13d07e0456157c26bbbee3bda

* Switch to workflow_call for GitHub Actions CI
- https://github.com/colcon/colcon-clean/commit/992e4b8ab7da5433fa74edfa8d3048d811db8f03

* Pin scantree to < 0.0.2a0
This version regresses support for Windows.

- https://github.com/colcon/colcon-clean/pull/46

* Fix package path for linter tests